Few improvements to the versioning scheme. First changed to a global monotonically increased version number (this time really like subversion and its repo-level version number). Second returning info about what has been really deployed in remote deployment (by web service) because if you deploy foo you will end up with a package like foo-12.
git-svn-id: https://svn.apache.org/repos/asf/incubator/ode/trunk@501522 13f79535-47bb-0310-9956-ffa450edef68

+++ bpel-store/src/main/java/org/apache/ode/store/DeploymentUnitDAO.java
@@ -51,7 +51,7 @@ public interface DeploymentUnitDAO {
      */
     void delete();
 
-    ProcessConfDAO createProcess(QName pid, QName type, int version);
+    ProcessConfDAO createProcess(QName pid, QName type, long version);
 
     ProcessConfDAO getProcess(QName pid);
